fix: pay product

+++ app/Http/Controllers/TransactionController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreTransactionRequest;
 use App\Http\Requests\UpdateTransactionRequest;
 use App\Models\Transaction;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Http;
 
 class TransactionController extends Controller
 {
@@ -81,7 +83,9 @@ class TransactionController extends Controller
      */
     public function destroy(Transaction $transaction)
     {
-        $transaction->delete();
+        if($transaction->status_id == 2){
+            $transaction->delete();
+        }
         return redirect()->route('client.history');
     }
 
@@ -94,5 +98,21 @@ class TransactionController extends Controller
             })
             ->where('store_id', auth()->user()->store->id)->get();
         return view('staff.sales',compact(['transactions']));
+    }
+
+    public function pay(Transaction $transaction, Request $request){
+        $payment_id =  $request->get('payment_id');
+
+        $response = Http::get("https://api.mercadopago.com/v1/payments/$payment_id"."?access_token=APP_USR-3157165296272046-110519-810a6e209a14209339eed5751121ab10-1537883018");
+
+        $response = json_decode($response);
+        $status = $response->status;
+
+        if($status == 'approved'){
+            $transaction->status_id = 1;
+            $transaction->save();
+        }
+
+        return redirect()->route('client.history');
     }
 }
